# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Money Guard",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Money Guard is a personal finance tracking application. It allows users to create, edit, and view transactions, analyze statistics, and check their current balance. The app uses secure authentication with JWT tokens."
  },
  "servers": [
    {
      "url": "https://moneyguard-app.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "description": "Register a new user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Anna"
                  },
                  "email": {
                    "type": "string",
                    "example": "anna@mail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "1QSET9xad"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "message",
                    "user"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User registered successfully"
                    },
                    "user": {
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/409"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "user": {
        "type": "object",
        "required": [
          "name",
          "email",
          "id"
        ],
        "properties": {
          "name": {
            "description": "User name",
            "type": "string",
            "example": "Anna"
          },
          "email": {
            "description": "User email",
            "type": "string",
            "example": "anna@mail.com"
          },
          "id": {
            "description": "User id",
            "type": "string",
            "example": "681a09af61bdc00bfe787f44"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
...
      "409": {
        "description": "User with this email already exist",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Email already in use"
                }
              }
            }
          }
        }
      },
...
    }
  }
}